Added more test cases to cover filtetring added

+++ Core/Pizza/Query/GetPizzasQuery.cs
@@ -23,7 +23,7 @@ public class GetPizzasQueryHandler(DatabaseContext databaseContext) : IRequestHa
         }
 
 
-        var entities = databaseContext.Pizzas.Select(x => x).AsNoTracking().FilterByDate(entity.DateCreated).FilterByName(entity.Name).FilterByPrice((float)entity.Price).OrderBy(entity.OrderBy);
+        var entities = databaseContext.Pizzas.Select(x => x).AsNoTracking().FilterByDate(entity.DateCreated).FilterByName(entity.Name).FilterByPrice(entity.Price).OrderBy(entity.OrderBy);
 
         var count = entities.Count();
         var paged = await entities.ApplyPaging(entity.PagingArgs).ToListAsync(cancellationToken);
